backend/requirements.txt
configparser
dotenv
flask
flask_cors
flask_mysqldb
mysql-connector-python
requests
python-dotenv
pyjwt
cryptography
